feat: move sqlalchemy models to dedicated module
