+++ models.py
@@ -1,0 +1,24 @@
+from sqlalchemy import Column, String, Integer, Boolean, DateTime, ForeignKey
+from sqlalchemy.orm import declarative_base, relationship
+import datetime
+
+Base = declarative_base()
+
+class User(Base):
+    __tablename__ = "users"
+    id = Column(Integer, primary_key=True)
+    email = Column(String, unique=True, nullable=False)
+    password_hash = Column(String, nullable=False)
+    phase = Column(String, default="explore")
+    is_verified = Column(Boolean, default=False)
+    created_at = Column(DateTime, default=datetime.datetime.utcnow)
+
+class EmailToken(Base):
+    __tablename__ = "email_tokens"
+    id = Column(Integer, primary_key=True)
+    user_id = Column(Integer, ForeignKey("users.id"))
+    token_hash = Column(String, unique=True, nullable=False)
+    type = Column(String)  # 'reset' or 'verify'
+    expires_at = Column(DateTime, nullable=False)
+    used = Column(Boolean, default=False)
+    user = relationship("User")
